# docs/09-api/openapi.yaml
openapi: 3.0.3
info:
  title: Ladushki Recipe Manager API
  version: 1.0.0
  description: REST API мобильного приложения для ведения кулинарных рецептов «Ладушки».
servers:
  - url: http://localhost:8082/api
    description: Docker backend
  - url: http://localhost:8080/api
    description: Local backend without Docker
  - url: http://10.0.2.2:8082/api
    description: Android emulator Docker backend alias
  - url: http://10.0.2.2:8080/api
    description: Android emulator local backend alias
security:
  - bearerAuth: []
...
paths:
...
  /complaints:
    post:
      tags: [Complaints]
      summary: Отправить жалобу на рецепт
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintRequest'
      responses:
        '200':
          description: Жалоба создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
...
  /admin/complaints/{id}/review:
    patch:
      tags: [Admin]
      summary: Взять жалобу в работу
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Жалоба взята в работу
  /admin/complaints/{id}/resolve:
    patch:
      tags: [Admin]
      summary: Закрыть жалобу
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Жалоба закрыта
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
          format: int64
        recipeId:
          type: integer
          format: int64
        recipeTitle:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Новая, В работе, Закрыта]
    ComplaintRequest:
      type: object
      required: [recipeId, recipeTitle, title, description]
      properties:
        recipeId:
          type: integer
          format: int64
        recipeTitle:
          type: string
        title:
          type: string
        description:
          type: string